+++ apps/client/public/CHANGELOG.md
@@ -1,6 +1,12 @@
 Changelog is only provided in English language.
 If you found some bugs or want us to add new feature, please do so via [our GitHub Tracker](https://github.com/sakilabs/openbts/issues/new) or send me an email [ririxi@sakilabs.com](mailto:ririxi@sakilabs.com)
 
+# 2026-04-04
+
+### 🚀 Enhancements
+
+- Cell analyzer now shows a **Band** column calculated from EARFCN (LTE) or UARFCN (UMTS)
+
 # 2026-04-03
 
 ### 🚀 Enhancements